Переделка моков комментариев
