+++ src/mock/comments.js
@@ -1,12 +1,62 @@
+import {nanoid} from 'nanoid';
 import {getRandomInteger} from './utils';
-import {generateComments} from './utils';
+
+/**
+ * Функция генерации комментария
+ * Определяет объект comment и случайным образом выбирает из него один элемент
+ * @return {object}
+ */
+export const generateComment = () => {
+  const comments = [
+    {
+      text: 'Interesting setting and a good cast',
+      author: 'Tim Macoveev',
+      emotion: 'smile',
+    },
+    {
+      text: 'Booooooooooring',
+      author: 'John Doe',
+      emotion: 'sleeping',
+    },
+    {
+      text: 'Very very old. Meh',
+      author: 'Elis',
+      emotion: 'puke',
+    },
+    {
+      text: 'Almost two hours? Seriously?',
+      author: 'Alex',
+      emotion: 'angry',
+    },
+    {
+      text: 'Very very old. Meh',
+      author: 'Suzan',
+      emotion: 'sleeping',
+    },
+  ];
+
+  const randomIndex = getRandomInteger(0, comments.length - 1);
+
+  return comments[randomIndex];
+};
+
+/**
+ * Функция создания объекта комментария
+ * Поля заполняются методами из этого файла
+ * @return {object}
+ */
+export const generateComments = () => {
+  return {
+    id: nanoid(),
+    info: generateComment(),
+    date: new Date(getRandomInteger(0, new Date())),
+  };
+};
 
 /**
  * Функция генерация времени + рандом
  * @return {string}
  */
-export const generateRandomComments = () => {
-  const COMMENT_COUNT = getRandomInteger(0, 5);
-  const randomComments = new Array(COMMENT_COUNT).fill(null).map((index) => generateComments(index));
-  return randomComments;
+export const generateRandomComments = (comment_count) => {
+  return new Array(comment_count).fill(null).map((index) => generateComments(index));
 };
