+++ src/mock/utils.js
@@ -16,58 +16,6 @@ export const getRandomInteger = (a = 0, b = 1) => {
 };
 
 /**
- * Функция генерации комментария
- * Определяет объект comment и случайным образом выбирает из него один элемент
- * @return {object}
- */
-export const generateComment = () => {
-  const comments = [
-    {
-      text: 'Interesting setting and a good cast',
-      author: 'Tim Macoveev',
-      emotion: 'smile',
-    },
-    {
-      text: 'Booooooooooring',
-      author: 'John Doe',
-      emotion: 'sleeping',
-    },
-    {
-      text: 'Very very old. Meh',
-      author: 'Elis',
-      emotion: 'puke',
-    },
-    {
-      text: 'Almost two hours? Seriously?',
-      author: 'Alex',
-      emotion: 'angry',
-    },
-    {
-      text: 'Very very old. Meh',
-      author: 'Suzan',
-      emotion: 'sleeping',
-    },
-  ];
-
-  const randomIndex = getRandomInteger(0, comments.length - 1);
-
-  return comments[randomIndex];
-};
-
-/**
- * Функция создания объекта комментария
- * Поля заполняются методами из этого файла
- * @return {object}
- */
-export const generateComments = () => {
-  return {
-    id: nanoid(),
-    info: generateComment(),
-    date: new Date(getRandomInteger(0, new Date())),
-  };
-};
-
-/**
  * Функция генерация рейтинга (рандомная генерация)
  * @param {number} a - начало промежутка
  * @param {number} b - конец промежутка
